refs #610 fix annotation support for parent classes hierarchy
